Update ArchiveFile.cs
support for rar5 archives.

You still need new 7z.dll:
Go to http://www.7-zip.org/download.html and get the newest x86 or x64 version and install it. Copy the "7z.dll" to the unrarit folder and rename it to "7z-amd64.dll". If you used the 32bit dll rename it to "7z-x86.dll"